# c1one/c1one-front/docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Uniqram Swagger",
    "description": "인스타그램 클론 프로젝트 API 명세서",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "JWT": []
    }
  ],
  "paths": {
...
    "/api/users/profiles": {
      "patch": {
        "tags": ["profile-controller"],
        "operationId": "patchProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/profiles/{userId}": {
      "get": {
        "tags": ["profile-controller"],
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProfileUpdateRequestDto": {
        "type": "object",
        "properties": {
          "bio": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "profileImageUrl": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bio": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "profileImageUrl": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}